Moved stages.java
